+++ src/main/java/im/aop/loggers/messageinterpolation/ObjectToStringStrategy.java
@@ -1,25 +1,20 @@
 package im.aop.loggers.messageinterpolation;
 
-import java.util.Arrays;
-import java.util.Objects;
-
 /**
- * {@link Objects#toString(Object)} implementation for {@link ToStringStrategy}.
+ * {@link ToStringStrategy} implementation using {@link String#valueOf(boolean)}. Return "null" if a
+ * is null.
  *
  * @author Andy Lian
  */
 public class ObjectToStringStrategy implements ToStringStrategy {
 
   @Override
-  public String toString(Object object) {
-    if (object == null) {
-      return "null";
-    }
+  public boolean canSupport(Class<?> type) {
+    return true;
+  }
 
-    if (object.getClass().isArray()) {
-      return Arrays.toString((Object[]) object);
-    } else {
-      return Objects.toString(object);
-    }
+  @Override
+  public String toString(Object object) {
+    return String.valueOf(object);
   }
 }