+++ src/main/java/im/aop/loggers/messageinterpolation/ToStringStrategy.java
@@ -8,5 +8,7 @@ package im.aop.loggers.messageinterpolation;
  */
 public interface ToStringStrategy {
 
+  boolean canSupport(Class<?> type);
+
   String toString(Object object);
 }