+++ src/main/java/im/aop/loggers/messageinterpolation/ToStringStrategyConfiguration.java
@@ -1,5 +1,7 @@
 package im.aop.loggers.messageinterpolation;
 
+import java.util.List;
+
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,8 +15,34 @@ import org.springframework.context.annotation.Configuration;
 class ToStringStrategyConfiguration {
 
   @Bean
-  @ConditionalOnMissingBean({ToStringStrategy.class})
-  public ObjectToStringStrategy toStringStrategy() {
+  @ConditionalOnMissingBean({ObjectToStringStrategy.class})
+  public ObjectToStringStrategy objectToStringStrategy() {
     return new ObjectToStringStrategy();
+  }
+
+  @Bean
+  public OptionalToStringStrategy optionalToStringStrategy(
+      final ObjectToStringStrategy objectToStringStrategy) {
+    return new OptionalToStringStrategy(objectToStringStrategy);
+  }
+
+  @Bean
+  public ArrayToStringStrategy arrayToStringStrategy(
+      final ObjectToStringStrategy objectToStringStrategy) {
+    return new ArrayToStringStrategy(objectToStringStrategy);
+  }
+
+  @Bean
+  public IterableToStringStrategy iterableToStringStrategy(
+      final ObjectToStringStrategy objectToStringStrategy) {
+    return new IterableToStringStrategy(objectToStringStrategy);
+  }
+
+  @Bean
+  @ConditionalOnMissingBean({ToStringStrategyFactory.class})
+  public ToStringStrategyFactory toStringStrategyFactory(
+      final ObjectToStringStrategy objectToStringStrategy,
+      final List<ToStringStrategy> toStringStrategies) {
+    return new DefaultToStringStrategyFactory(objectToStringStrategy, toStringStrategies);
   }
 }